pyproject.toml
[project]
name = "analytics"
version = "0.1.0"
description = ""
authors = [
    {name = "Shubham Bhardwaj",email = "Shubbi.30nov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python (>=2.1.5,<3.0.0)",
    "pyspark (>=3.5.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
